clerk

+++ lib/actions/user.action.ts
@@ -4,8 +4,8 @@ import User from "../models/user.model";
 import { connectToDatabase } from "../dbconnection";
 import Event from "../models/event.model";
 import Order from "../models/order.model";
+import { redirect } from "next/dist/server/api-utils";
 import { revalidatePath } from "next/cache";
-
 export interface CreateUserParams {
 	clerkId: string;
 	email: string;
@@ -18,10 +18,12 @@ export interface CreateUserParams {
 export async function createUser(userData: CreateUserParams) {
 	try {
 		await connectToDatabase();
+
 		const user = await User.create(userData);
+
 		return JSON.parse(JSON.stringify(user));
 	} catch (error) {
-		console.error("Error creating user:", error);
+		console.log(error);
 		throw error;
 	}
 }
@@ -29,19 +31,16 @@ export async function createUser(userData: CreateUserParams) {
 export async function getUserByClerkId(clerkId: string) {
 	try {
 		await connectToDatabase();
-		console.log(`Fetching user with clerkId: ${clerkId}`);
 
 		const user = await User.findOne({ clerkId: clerkId });
 
 		if (!user) {
-			console.error(`User with clerkId ${clerkId} not found`);
 			throw new Error("User not found");
 		}
 
-		console.log(`User found: ${JSON.stringify(user)}`);
 		return JSON.parse(JSON.stringify(user));
 	} catch (error) {
-		console.error(`Error fetching user by clerkId ${clerkId}:`, error);
+		console.log(error);
 		throw error;
 	}
 }
@@ -49,14 +48,16 @@ export async function getUserByClerkId(clerkId: string) {
 export async function getUserById(userId: string) {
 	try {
 		await connectToDatabase();
+
 		const user = await User.findById(userId);
+
 		if (!user) {
-			console.error(`User with ID ${userId} not found`);
 			throw new Error("User not found");
 		}
+
 		return JSON.parse(JSON.stringify(user));
 	} catch (error) {
-		console.error(`Error fetching user by ID ${userId}:`, error);
+		console.log(error);
 		throw error;
 	}
 }
@@ -74,21 +75,19 @@ export interface UpdateUserParams {
 export async function updateUser(params: UpdateUserParams) {
 	try {
 		await connectToDatabase();
+
 		const user = await User.findOneAndUpdate(
 			{ clerkId: params.clerkId },
-			params.userData,
-			{ new: true } // This option returns the updated document
+			params.userData
 		);
+
 		if (!user) {
-			console.error(`User with clerkId ${params.clerkId} not found`);
 			throw new Error("User not found");
 		}
+
 		return JSON.parse(JSON.stringify(user));
 	} catch (error) {
-		console.error(
-			`Error updating user with clerkId ${params.clerkId}:`,
-			error
-		);
+		console.log(error);
 		throw error;
 	}
 }
@@ -96,17 +95,30 @@ export async function updateUser(params: UpdateUserParams) {
 export async function deleteUser(clerkId: string) {
 	try {
 		await connectToDatabase();
+
 		const user = await User.findOne({ clerkId: clerkId });
+
 		if (!user) {
-			console.error(`User with clerkId ${clerkId} not found`);
 			throw new Error("User not found");
 		}
-		await Event.deleteMany({ organizer: user._id });
-		await Order.deleteMany({ user: user._id });
+
+		const userEvents = await Event.find({ organizer: user._id });
+
+		if (userEvents) {
+			await Event.deleteMany({ organizer: user._id });
+		}
+
+		const userOrders = await Order.find({ user: user._id });
+
+		if (userOrders) {
+			await Order.deleteMany({ user: user._id });
+		}
+
 		await User.findByIdAndDelete(user._id);
+
 		return JSON.parse(JSON.stringify(user));
 	} catch (error) {
-		console.error(`Error deleting user with clerkId ${clerkId}:`, error);
+		console.log(error);
 		throw error;
 	}
 }
@@ -114,20 +126,24 @@ export async function deleteUser(clerkId: string) {
 export async function likeEvent(eventId: string, userId: string) {
 	try {
 		await connectToDatabase();
+
 		const event = await Event.findById(eventId);
+
 		const user = await User.findById(userId);
+
 		if (!event) {
-			console.error(`Event with ID ${eventId} not found`);
 			throw new Error("Event not found");
 		}
+
 		if (!user) {
-			console.error(`User with ID ${userId} not found`);
 			throw new Error("Please login to like an event");
 		}
+
 		const alreadyLiked = await User.findOne({
 			_id: user._id,
 			likedEvents: eventId,
 		});
+
 		if (!alreadyLiked) {
 			await User.findByIdAndUpdate(user._id, {
 				$push: { likedEvents: eventId },
@@ -137,14 +153,17 @@ export async function likeEvent(eventId: string, userId: string) {
 				$pull: { likedEvents: eventId },
 			});
 		}
+
 		revalidatePath(`/`);
 		revalidatePath(`/likes`);
-		return !!alreadyLiked;
+
+		if (alreadyLiked) {
+			return true;
+		} else {
+			return false;
+		}
 	} catch (error) {
-		console.error(
-			`Error liking event ${eventId} by user ${userId}:`,
-			error
-		);
+		console.log(error);
 		throw error;
 	}
 }
@@ -152,11 +171,13 @@ export async function likeEvent(eventId: string, userId: string) {
 export async function getLikedEvents(userId: string) {
 	try {
 		await connectToDatabase();
+
 		const user = await User.findById(userId);
+
 		if (!user) {
-			console.error(`User with ID ${userId} not found`);
 			throw new Error("User not found");
 		}
+
 		const response = await User.findById(userId).populate({
 			path: "likedEvents",
 			populate: [
@@ -164,9 +185,10 @@ export async function getLikedEvents(userId: string) {
 				{ path: "category", model: "Category" },
 			],
 		});
+
 		return JSON.parse(JSON.stringify(response.likedEvents));
 	} catch (error) {
-		console.error(`Error fetching liked events for user ${userId}:`, error);
+		console.log(error);
 		throw error;
 	}
 }
